third_party/blink/web_tests/http/tests/security/popup-allowed-by-sandbox-is-sandboxed.html

<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
    testRunner.setPopupBlockingEnabled(false);
}
</script>
<p>To run this test outside of DumpRenderTree, please disable your popup blocker!</p>
<p>If you change this test, please be sure to change popup-allowed-by-sandbox-is-sandboxed-control.html as well!</p>
<p>This test passes if it doesn't alert FAIL.</p>
<iframe sandbox="allow-scripts allow-popups allow-same-origin"
  src="data:text/html,
       <script>
       var win = window.open('about:blank', '_blank');
       win.document.write(
          '<form action=javascript:alert(/FAIL/) ><input type=submit></form>' +
          '<script>document.forms[0].submit();' +
          'if (window.testRunner)testRunner.notifyDone();' +
          '</' + 'script>');
       win.document.close();
       </script>"
  ></iframe>
