third_party/blink/web_tests/http/tests/security/popup-allowed-by-sandbox-is-sandboxed-expected.txt

CONSOLE WARNING: An iframe which has both allow-scripts and allow-same-origin for its sandbox attribute can escape its sandboxing.
CONSOLE ERROR: Blocked form submission to 'javascript:alert(/FAIL/)' because the form's frame is sandboxed and the 'allow-forms' permission is not set.
To run this test outside of DumpRenderTree, please disable your popup blocker!

If you change this test, please be sure to change popup-allowed-by-sandbox-is-sandboxed-control.html as well!

This test passes if it doesn't alert FAIL.
